feat: Remove GPT integration and rename operations to tasks
