+++ packages/pinion/package.json
@@ -54,9 +54,7 @@
     "access": "public"
   },
   "dependencies": {
-    "@feathershq/api": "^0.3.1",
     "@types/inquirer": "^9.0.7",
-    "chalk": "^4.0.0",
     "commander": "^11.1.0",
     "inquirer": "^8.0.0",
     "open": "^8.4.0",
@@ -67,7 +65,6 @@
     "@types/node": "^20.10.6",
     "mocha": "^10.2.0",
     "shx": "^0.3.4",
-    "ts-node": "^10.9.2",
     "typescript": "^5.3.3"
   },
   "gitHead": "ea97aee4f8bc2b6ca1a78c5bc3dd686c1a57d5c8"
