--- packages/pinion/src/gpt/index.ts
@@ -1,2 +1,0 @@
-export * from './operation'
-export * from './client'
